feat: slugify filter and modularization
fix #48
