+++ example/templates/content.html
@@ -16,7 +16,7 @@
         <span class="content-date"> {{ content.date | date(format="%v") }}</span>
         <ul class="content-tags">
         {% for tag in content.tags  %}
-            <li><a href="{{url_for(path='/tag/')}}{{ tag | trim }}.html">{{ tag }}</a></li>
+            <li><a href="{{url_for(path='/tag/')}}{{ tag | trim | slugify }}.html">{{ tag }}</a></li>
         {% endfor %}
         </ul>
     </footer>
