+++ example/templates/group.html
@@ -6,7 +6,7 @@
         <article class="group-list">
         <ul class="content-tags">
             {% for content in group_content -%}
-                <li><a href="{{url_for(path='/tag/')}}{{ content.0 | trim }}.html">{{ content.0 | capitalize }}</a> [{{content.1}}]</li>
+                <li><a href="{{url_for(path='/tag/')}}{{ content.0 | trim | slugify }}.html">{{ content.0 | capitalize }}</a> [{{content.1}}]</li>
             {%- endfor %}
         </ul>
         </article>
