+++ example/templates/list.html
@@ -16,7 +16,7 @@
                     {% if content.tags -%}
                     <ul class="content-tags">
                         {% for tag in content.tags | slice(end=3) -%}
-                            <li><a href="{{url_for(path='/tag/')}}{{ tag | trim }}.html">{{ tag }}</a></li>
+                            <li><a href="{{url_for(path='/tag/')}}{{ tag | trim | slugify }}.html">{{ tag }}</a></li>
                         {%- endfor %}
                     </ul>
                     {%- endif %}
